# docs/swagger.yaml
definitions:
...
  subscriptions.GetAllResponse:
    properties:
      errors:
        description: Error messages, present only on failure responses
        example:
        - invalid url
        items:
          type: string
        type: array
      subscriptions:
        items:
          $ref: '#/definitions/subscriptions.SubscriptionInfo'
        type: array
    type: object
  subscriptions.SubscribeRequest:
    properties:
      event_types:
        description: Event types to subscribe the endpoint to
        example:
        - order.created
        items:
          type: string
        type: array
    required:
    - event_types
    type: object
  subscriptions.SubscribeResponse:
    properties:
      endpoint_id:
        example: ep_1a2b3c
        type: string
      errors:
        description: Error messages, present only on failure responses
        example:
        - invalid url
        items:
          type: string
        type: array
      subscriptions:
        items:
          $ref: '#/definitions/subscriptions.SubscriptionInfo'
        type: array
    type: object
  subscriptions.SubscriptionInfo:
    properties:
      created_at:
        example: "2026-07-01T12:00:00Z"
        type: string
      endpoint_id:
        example: ep_1a2b3c
        type: string
      event_type:
        example: order.created
        type: string
      id:
        example: sub_1a2b3c
        type: string
    type: object
  utils.Response:
    properties:
      errors:
        description: Error messages, present only on failure responses
        example:
        - invalid url
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: An open-source webhook delivery service in Go
  title: Webhook Delivery
  version: 1.0.0
paths:
...
  /endpoints/{id}/subscriptions:
    get:
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.GetAllResponse'
        "400":
          description: id is not provided
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: endpoint not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List an endpoint's subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Adds subscriptions for the given event types. Idempotent — event
        types the endpoint is already subscribed to are ignored.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: string
      - description: Event types to subscribe to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscriptions.SubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.SubscribeResponse'
        "400":
          description: id or event types not provided
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: endpoint not found
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: subscription already exists
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Subscribe an endpoint to event types
      tags:
      - subscriptions
...
swagger: "2.0"